Support decimal data types, for arbitrary precision
The import/export code currently uses `decimal` as the type for SQL
Server's decimal columns. This may lead to a loss of precision in some
cases, but makes it simpler to handle `money` columns cleanly.

The underlying file format will handle any precision perfectly. This is
solely an implementation limitation.